test(workflow-core): add dedicated GlobalPortIdentitySerde spec (#5717)
Add a dedicated `GlobalPortIdentitySerdeSpec` for `GlobalPortIdentitySerde`,
the helper that serializes/deserializes a `GlobalPortIdentity` to the compact,
underscore-free string used in VFS URIs and file paths.

`GlobalPortIdentitySerde` was previously exercised inside
`PortIdentitySerdeSpec`, which mixed it with the unrelated PortIdentity Jackson
key (de)serializer tests. Move those cases into their own file (the name issue
#5717 asks for) so the two serde concerns are tested independently, and trim
`PortIdentitySerdeSpec` down to just the Jackson key (de)serialization tests
(dropping the now-unused imports). No production code changes; no duplication
between the two files.

The new spec keeps the existing round-trip / format / validation coverage and
adds four edge cases that were previously unpinned:

- serialize/deserialize asymmetry: the deserializer regex accepts underscores
  even though the serializer rejects them (characterized so a future tightening
  breaks the test on purpose)
- unescaped comma in logicalOpId breaks the round-trip (the format does not
  escape its `,` separator)
- portId boundary: Int.MaxValue round-trips, and an out-of-Int-range value is
  rejected with NumberFormatException on deserialize
- mixed-case booleans (`True` / `FALSE`) are accepted on deserialize, since
  Scala's String.toBoolean is case-insensitive

Verified with:
  sbt "WorkflowCore/testOnly org.apache.texera.amber.util.serde.GlobalPortIdentitySerdeSpec org.apache.texera.amber.util.serde.PortIdentitySerdeSpec"
-> 28 tests, 0 failures (1 pendingUntilFixed, pre-existing).
